Update the AEA3 reading compliant to the datasheet (1x16 consecutives clock cycles)

+++ emBODY/eBcode/arch-arm/embobj/plus/mc/AbsEncoder.c
@@ -491,7 +491,7 @@ void AbsEncoder_update(AbsEncoder* o, uint16_t position)
     
     o->delta = 0;
     
-    int16_t check = position - o->position_last;
+    int32_t check = position - o->position_last;
     
     o->position_last = position;
